1-3088: extract project health grid
