+++ frontend/src/component/project/Project/ProjectStatus/ProjectStatusModal.tsx
@@ -1,13 +1,11 @@
 import { styled } from '@mui/material';
 import { DynamicSidebarModal } from 'component/common/SidebarModal/SidebarModal';
 import { ReactComponent as ProjectStatusSvg } from 'assets/icons/projectStatus.svg';
-import { ProjectResources } from './ProjectResources';
 import { ProjectActivity } from './ProjectActivity';
-import { ProjectHealth } from './ProjectHealth';
 import { ProjectLifecycleSummary } from './ProjectLifecycleSummary';
-import { StaleFlags } from './StaleFlags';
 import type { FC } from 'react';
 import { HelpIcon } from 'component/common/HelpIcon/HelpIcon';
+import { ProjectHealthGrid } from './ProjectHealthGrid';
 
 const ModalContentContainer = styled('section')(({ theme }) => ({
     minHeight: '100vh',
@@ -25,32 +23,6 @@ type Props = {
     open: boolean;
     close: () => void;
 };
-
-const onNarrowGrid = (css: object) => ({
-    '@container (max-width: 650px)': css,
-    '@supports not (container-type: inline-size)': {
-        '@media (max-width: 712px)': css,
-    },
-});
-
-const HealthContainer = styled('div')({
-    containerType: 'inline-size',
-});
-
-const HealthGrid = styled('div')(({ theme }) => ({
-    display: 'grid',
-    gridTemplateAreas: `
-        "health resources"
-        "stale resources"
-    `,
-    gridTemplateColumns: '1fr 1fr',
-    gap: theme.spacing(1, 2),
-    ...onNarrowGrid({
-        display: 'flex',
-        flexDirection: 'column',
-        gap: theme.spacing(1),
-    }),
-}));
 
 const LifecycleHeaderRow = styled('div')(({ theme }) => ({
     display: 'flex',
@@ -130,13 +102,7 @@ export const ProjectStatusModal = ({ open, close }: Props) => {
                 </HeaderRow>
                 <Row>
                     <RowHeader>Health</RowHeader>
-                    <HealthContainer>
-                        <HealthGrid>
-                            <ProjectHealth />
-                            <StaleFlags />
-                            <ProjectResources />
-                        </HealthGrid>
-                    </HealthContainer>
+                    <ProjectHealthGrid />
                 </Row>
 
                 <Row>
